refactor: split components into individual modules - use pascal case for component files
