+++ src/index.ts
@@ -1,0 +1,10 @@
+export { ReactCompareSlider } from './ReactCompareSlider';
+export type { ReactCompareSliderProps } from './ReactCompareSlider';
+
+export { ReactCompareSliderHandle } from './ReactCompareSliderHandle';
+export type { ReactCompareSliderHandleProps } from './ReactCompareSliderHandle';
+
+export { ReactCompareSliderImage } from './ReactCompareSliderImage';
+export type { ReactCompareSliderImageProps } from './ReactCompareSliderImage';
+
+export { styleFitContainer } from './utils';
