--- src/index.tsx
@@ -1,13 +1,0 @@
-export {
-  ReactCompareSlider,
-  ReactCompareSliderProps,
-  ReactCompareSliderHandle,
-  ReactCompareSliderHandleProps,
-} from './react-compare-slider';
-
-export {
-  ReactCompareSliderImage,
-  ReactCompareSliderImageProps,
-} from './react-compare-slider-image';
-
-export { styleFitContainer } from './utils';
