+++ test/ReactCompareSliderImage.test.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import { render, cleanup } from '@testing-library/react';
-import '@testing-library/jest-dom/extend-expect';
 
 import { ReactCompareSliderImage } from '../src';
